Added CurriedCSVReader and moved MatchingFiles to pipeline.util.
