+++ pipeline/util/__init__.py
@@ -1,10 +1,12 @@
+import os
+import fnmatch
 from threading import Lock
 from contextlib import ContextDecorator, suppress
 from collections import defaultdict
 
 import settings
 import pipeline.io.arches
-from bonobo.config import Configurable, Option
+from bonobo.config import Configurable, Option, Service
 from cromulent.model import factory
 
 def identity(d):
@@ -140,3 +142,25 @@ class ExtractKeyedValues(Configurable):
 				'parent_data': data,
 			})
 			yield child
+
+class MatchingFiles(Configurable):
+	'''
+	Given a path and a pattern, yield the names of all files in the path that match the pattern.
+	'''
+	path = Option(str)
+	pattern = Option(str, default='*')
+	fs = Service(
+		'fs',
+		__doc__='''The filesystem instance to use.''',
+	)  # type: str
+	def __call__(self, *, fs, **kwargs):
+		count = 0
+		subpath, pattern = os.path.split(self.pattern)
+		fullpath = os.path.join(self.path, subpath)
+		for f in sorted(fs.listdir(fullpath)):
+			if fnmatch.fnmatch(f, pattern):
+				yield os.path.join(subpath, f)
+				count += 1
+		if not count:
+			sys.stderr.write(f'*** No files matching {pattern} found in {fullpath}\n')
+
